requirements.txt
streamlit
pandas
numpy
plotly
yfinance
scikit-learn
python-dotenv
requests
groq
scipy
beautifulsoup4
